BUG/MINOR: base64: base64urldec() ignores padding in output size check
Without this fix, the decode function would proceed even when the output
buffer is not large enough, because the padding was not considered. For
example, it would not fail with the input length of 23 and the output
buffer size of 15, even the actual decoded output size is 17.

This patch should be backported to all stable branches that have a
base64urldec() function available.

(cherry picked from commit 61aa4428c1a0a7b747914da0f7b47bae59f4f755)
Signed-off-by: Willy Tarreau <w@1wt.eu>