BUG/MINOR: vars: do not talk about global section in CLI errors for set-var
When attempting to set a variable does not start with the "proc" scope on
the CLI, we used to emit "only proc is permitted in the global section"
which obviously is a leftover from the initial code.

This may be backported to 2.4.

(cherry picked from commit c767eebf1f2130447cef3269f4a46ed164410a75)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>