BUG/MAJOR: mux-h1: Don't eval input data if an error was reported
If an error was already reported on the H1 connection, pending input data
must not be (re)evaluated in h1_process(). Otherwise an unexpected internal
error will be reported, in addition of the first one. And on some
conditions, this may generate an infinite loop because the mux tries to send
an internal error but it fails to do so thus it loops to retry.

This patch should fix the issue #1356. It must be backported to 2.4.

(cherry picked from commit ab7389dc3cef3bf09ef23a1127d4bc0ca7a8b0a0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>