BUG/MEDIUM: http: check for a channel pending data before waiting
If there's no channel input data, we should still check for possible
pending data starting at chn->buf.p and try to forward it. Without this
patch, in some cases, it is possible to end up waiting for a long time.

The issue was introduced with commit ca874b8d9 ("BUG/MEDIUM: http: Don't
parse chunked body if there is no input data").

No need to backport, this is 2.0-only.