DOC: peers: fix doc "enable" statement on "peers" sections
Checking in code the right keyword is "enabled" and not "enable".

In addition the comment was also completed:

This could appear useless because the "defaults" sections not
yet apply on "peers" sections, but it could be the case in the future.
This statement can currently cancel a previous "disabled" keyword in
the same section.

This patch should be backported in all supported branches (keyword
is present since 1.5)

(cherry picked from commit 620761f934fe24f3943ec2ff346d8ecd0e6a6576)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>